+++ NEWS.md
@@ -7,12 +7,13 @@
 
 ## New functions / features
 
+- `point_estimate()` added function to compute point estimates
 - `p_direction()`: new argument `method` to compute pd based on AUC
 - `area_under_curve()`: compute AUC
 - `distribution()` functions have been added
 - `bayesfactor()` function has been added
 - Started adding plotting methods (currently in the [`see`](https://github.com/easystats/see) package) for `p_direction()` and `hdi()`
-- `probability_at()` as alias for `density_at()`.
+- `probability_at()` as alias for `density_at()`
 
 ## Minor changes